feat: [DEL-3944]: Rename PerpetualTask client file
