+++ 260-delegate/src/main/java/io/harness/perpetualtask/PerpetualTaskLifecycleManager.java
@@ -15,7 +15,6 @@ import io.harness.concurrent.HTimeLimiter;
 import io.harness.grpc.utils.AnyUtils;
 import io.harness.grpc.utils.HTimestamps;
 import io.harness.logging.AutoLogContext;
-import io.harness.perpetualtask.grpc.PerpetualTaskServiceGrpcClient;
 
 import com.google.common.util.concurrent.TimeLimiter;
 import com.google.common.util.concurrent.UncheckedTimeoutException;
@@ -36,16 +35,16 @@ public class PerpetualTaskLifecycleManager {
   private final PerpetualTaskExecutionParams params;
   private final PerpetualTaskExecutionContext context;
   private final PerpetualTaskExecutor perpetualTaskExecutor;
-  private final PerpetualTaskServiceGrpcClient perpetualTaskServiceGrpcClient;
+  private final PerpetualTaskServiceAgentClient perpetualTaskServiceAgentClient;
   private final AtomicInteger currentlyExecutingPerpetualTasksCount;
 
   PerpetualTaskLifecycleManager(PerpetualTaskId taskId, PerpetualTaskExecutionContext context,
-      Map<String, PerpetualTaskExecutor> factoryMap, PerpetualTaskServiceGrpcClient perpetualTaskServiceGrpcClient,
+      Map<String, PerpetualTaskExecutor> factoryMap, PerpetualTaskServiceAgentClient perpetualTaskServiceAgentClient,
       TimeLimiter timeLimiter, AtomicInteger currentlyExecutingPerpetualTasksCount) {
     this.taskId = taskId;
     this.context = context;
     this.timeLimiter = timeLimiter;
-    this.perpetualTaskServiceGrpcClient = perpetualTaskServiceGrpcClient;
+    this.perpetualTaskServiceAgentClient = perpetualTaskServiceAgentClient;
     params = context.getTaskParams();
     perpetualTaskExecutor = factoryMap.get(getTaskType(params));
     timeoutMillis = Durations.toMillis(context.getTaskSchedule().getTimeout());
@@ -92,7 +91,7 @@ public class PerpetualTaskLifecycleManager {
       log.error("Exception is ", ex);
       decrementTaskCounter();
     }
-    perpetualTaskServiceGrpcClient.heartbeat(taskId, taskStartTime, perpetualTaskResponse);
+    perpetualTaskServiceAgentClient.heartbeat(taskId, taskStartTime, perpetualTaskResponse);
     return null;
   }
