+++ 260-delegate/src/main/java/io/harness/perpetualtask/PerpetualTaskWorker.java
@@ -19,7 +19,6 @@ import io.harness.flow.BackoffScheduler;
 import io.harness.logging.AutoLogContext;
 import io.harness.logging.LoggingListener;
 import io.harness.mongo.DelayLogContext;
-import io.harness.perpetualtask.grpc.PerpetualTaskServiceGrpcClient;
 import io.harness.threading.Schedulable;
 
 import com.google.common.annotations.VisibleForTesting;
@@ -66,7 +65,7 @@ public class PerpetualTaskWorker {
 
   private final AtomicBoolean firstFillUp = new AtomicBoolean(true);
   private final BackoffScheduler backoffScheduler;
-  private final PerpetualTaskServiceGrpcClient perpetualTaskServiceGrpcClient;
+  private final PerpetualTaskServiceAgentClient perpetualTaskServiceAgentClient;
   private final Map<String, PerpetualTaskExecutor> factoryMap;
 
   private final AtomicBoolean running = new AtomicBoolean(false);
@@ -91,10 +90,10 @@ public class PerpetualTaskWorker {
   }
 
   @Inject
-  public PerpetualTaskWorker(PerpetualTaskServiceGrpcClient perpetualTaskServiceGrpcClient,
+  public PerpetualTaskWorker(PerpetualTaskServiceAgentClient perpetualTaskServiceAgentClient,
       Map<String, PerpetualTaskExecutor> factoryMap, @Named("taskExecutor") ThreadPoolExecutor perpetualTaskExecutor,
       @Named("perpetualTaskTimeoutExecutor") ScheduledExecutorService perpetualTaskTimeoutExecutor) {
-    this.perpetualTaskServiceGrpcClient = perpetualTaskServiceGrpcClient;
+    this.perpetualTaskServiceAgentClient = perpetualTaskServiceAgentClient;
     this.factoryMap = factoryMap;
     this.perpetualTaskTimeLimiter = HTimeLimiter.create(perpetualTaskExecutor);
     this.perpetualTaskTimeoutExecutor = perpetualTaskTimeoutExecutor;
@@ -182,7 +181,7 @@ public class PerpetualTaskWorker {
 
   List<PerpetualTaskAssignDetails> fetchAssignedTask() {
     String delegateId = getDelegateId().orElse("UNREGISTERED");
-    List<PerpetualTaskAssignDetails> assignedTasks = perpetualTaskServiceGrpcClient.perpetualTaskList(delegateId);
+    List<PerpetualTaskAssignDetails> assignedTasks = perpetualTaskServiceAgentClient.perpetualTaskList(delegateId);
     if (log.isDebugEnabled()) {
       log.debug("Refreshed list of assigned perpetual tasks {}", assignedTasks);
     }
@@ -192,12 +191,12 @@ public class PerpetualTaskWorker {
   @VisibleForTesting
   void startTask(PerpetualTaskAssignDetails task) {
     try (AutoLogContext ignore1 = new PerpetualTaskLogContext(task.getTaskId().getId(), OVERRIDE_ERROR)) {
-      PerpetualTaskExecutionContext context = perpetualTaskServiceGrpcClient.perpetualTaskContext(task.getTaskId());
+      PerpetualTaskExecutionContext context = perpetualTaskServiceAgentClient.perpetualTaskContext(task.getTaskId());
       PerpetualTaskSchedule schedule = context.getTaskSchedule();
       long intervalSeconds = Durations.toSeconds(schedule.getInterval());
 
       PerpetualTaskLifecycleManager perpetualTaskLifecycleManager =
-          new PerpetualTaskLifecycleManager(task.getTaskId(), context, factoryMap, perpetualTaskServiceGrpcClient,
+          new PerpetualTaskLifecycleManager(task.getTaskId(), context, factoryMap, perpetualTaskServiceAgentClient,
               perpetualTaskTimeLimiter, currentlyExecutingPerpetualTasksCount);
 
       synchronized (runningTaskMap) {
