+++ 260-delegate/src/test/java/io/harness/perpetualtask/PerpetualTaskWorkerTest.java
@@ -19,7 +19,6 @@ import static org.mockito.Mockito.when;
 import io.harness.DelegateTestBase;
 import io.harness.category.element.UnitTests;
 import io.harness.manage.ManagedScheduledExecutorService;
-import io.harness.perpetualtask.grpc.PerpetualTaskServiceGrpcClient;
 import io.harness.perpetualtask.k8s.watch.K8sWatchTaskParams;
 import io.harness.rule.Owner;
 import io.harness.serializer.KryoSerializer;
@@ -72,7 +71,7 @@ public class PerpetualTaskWorkerTest extends DelegateTestBase {
   PerpetualTaskExecutionContext context;
 
   @Rule public MockitoRule mockitoRule = MockitoJUnit.rule();
-  @Mock private PerpetualTaskServiceGrpcClient perpetualTaskServiceGrpcClient;
+  @Mock private PerpetualTaskServiceAgentClient perpetualTaskServiceAgentClient;
   @Mock private Map<String, PerpetualTaskExecutor> factoryMap;
   @Mock @Named("taskExecutor") ThreadPoolExecutor perpetualTaskExecutor;
   @Spy
@@ -95,7 +94,7 @@ public class PerpetualTaskWorkerTest extends DelegateTestBase {
     PerpetualTaskSchedule schedule = PerpetualTaskSchedule.newBuilder().setInterval(Durations.fromSeconds(1)).build();
     context = PerpetualTaskExecutionContext.newBuilder().setTaskParams(params).setTaskSchedule(schedule).build();
 
-    when(perpetualTaskServiceGrpcClient.perpetualTaskContext(isA(PerpetualTaskId.class))).thenReturn(context);
+    when(perpetualTaskServiceAgentClient.perpetualTaskContext(isA(PerpetualTaskId.class))).thenReturn(context);
   }
 
   @Test
@@ -127,7 +126,7 @@ public class PerpetualTaskWorkerTest extends DelegateTestBase {
   @Category(UnitTests.class)
   public void testFetchAssignedTask() {
     worker.fetchAssignedTask();
-    verify(perpetualTaskServiceGrpcClient).perpetualTaskList(anyString());
+    verify(perpetualTaskServiceAgentClient).perpetualTaskList(anyString());
   }
 
   @Test
